# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elite FF Tournament Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  role: string;
}

export interface Tournament {
  id: number;
  name: string;
  type: string;
  mode: string;
  /** @nullable */
  mapName?: string | null;
  teamSize: string;
  /** @nullable */
  entryFee?: number | null;
  /** @nullable */
  prizePool?: number | null;
  /** @nullable */
  booyahPrize?: number | null;
  /** @nullable */
  secondPrize?: number | null;
  /** @nullable */
  thirdPrize?: number | null;
  /** @nullable */
  highestKillPrize?: number | null;
  maxSlots: number;
  filledSlots: number;
  status: string;
  scheduledAt: string;
  /** @nullable */
  rules?: string | null;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  qrUrl?: string | null;
  /** @nullable */
  roomId?: string | null;
  /** @nullable */
  roomPassword?: string | null;
  /** @nullable */
  cancelReason?: string | null;
  /** @nullable */
  delayInfo?: string | null;
  isPaid: boolean;
  timerEnabled: boolean;
  createdAt: string;
  hostId: number;
}

export interface CreateTournamentBody {
  name: string;
  type: string;
  mode: string;
  /** @nullable */
  mapName?: string | null;
  teamSize: string;
  /** @nullable */
  entryFee?: number | null;
  /** @nullable */
  prizePool?: number | null;
  /** @nullable */
  booyahPrize?: number | null;
  /** @nullable */
  secondPrize?: number | null;
  /** @nullable */
  thirdPrize?: number | null;
  /** @nullable */
  highestKillPrize?: number | null;
  maxSlots: number;
  scheduledAt: string;
  /** @nullable */
  rules?: string | null;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  upiId?: string | null;
  isPaid: boolean;
  timerEnabled: boolean;
}

export interface UpdateTournamentBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  mode?: string | null;
  /** @nullable */
  mapName?: string | null;
  /** @nullable */
  teamSize?: string | null;
  /** @nullable */
  entryFee?: number | null;
  /** @nullable */
  prizePool?: number | null;
  /** @nullable */
  booyahPrize?: number | null;
  /** @nullable */
  secondPrize?: number | null;
  /** @nullable */
  thirdPrize?: number | null;
  /** @nullable */
  highestKillPrize?: number | null;
  /** @nullable */
  maxSlots?: number | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  rules?: string | null;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  isPaid?: boolean | null;
  /** @nullable */
  timerEnabled?: boolean | null;
  /** @nullable */
  roomId?: string | null;
  /** @nullable */
  roomPassword?: string | null;
}

export interface CancelTournamentBody {
  reason: string;
}

export interface DelayTournamentBody {
  newScheduledAt: string;
  reason: string;
}

export interface RoomCredentialsBody {
  roomId: string;
  roomPassword: string;
}

export interface Registration {
  id: number;
  tournamentId: number;
  userId: number;
  squadName: string;
  playerNames: string;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  /** @nullable */
  upiId?: string | null;
  utrNumber: string;
  status: string;
  /** @nullable */
  slotNumber?: number | null;
  /** @nullable */
  declineReason?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  createdAt: string;
  user: User;
}

export interface CreateRegistrationBody {
  tournamentId: number;
  squadName: string;
  playerNames: string;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  /** @nullable */
  upiId?: string | null;
  utrNumber: string;
}

export interface VerifyRegistrationBody {
  /** @nullable */
  slotNumber?: number | null;
}

export interface DeclineRegistrationBody {
  reason: string;
}

export interface MatchResult {
  id: number;
  tournamentId: number;
  registrationId: number;
  squadName: string;
  /** @nullable */
  placement?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  kills?: number | null;
  /** @nullable */
  prize?: number | null;
  /** @nullable */
  prizeType?: string | null;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  /** @nullable */
  utrNumber?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CreateResultBody {
  tournamentId: number;
  registrationId: number;
  squadName: string;
  /** @nullable */
  placement?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  kills?: number | null;
  /** @nullable */
  prize?: number | null;
  /** @nullable */
  prizeType?: string | null;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  /** @nullable */
  utrNumber?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface UpdateResultBody {
  /** @nullable */
  placement?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  kills?: number | null;
  /** @nullable */
  prize?: number | null;
  /** @nullable */
  prizeType?: string | null;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  /** @nullable */
  utrNumber?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  type: string;
  /** @nullable */
  tournamentId?: number | null;
  /** @nullable */
  registrationId?: number | null;
  isRead: boolean;
  createdAt: string;
}

export interface CreateNotificationBody {
  /** @nullable */
  userId?: number | null;
  title: string;
  message: string;
  type: string;
  /** @nullable */
  tournamentId?: number | null;
  /** @nullable */
  registrationId?: number | null;
  targetAll?: boolean;
}

export interface Feedback {
  id: number;
  /** @nullable */
  userId?: number | null;
  name: string;
  message: string;
  /** @nullable */
  rating?: number | null;
  createdAt: string;
}

export interface FeedbackBody {
  name: string;
  message: string;
  /** @nullable */
  rating?: number | null;
}

export interface ContactBody {
  name: string;
  email?: string;
  message: string;
  via: string;
}

export interface StatsSummary {
  totalTournaments: number;
  activeTournaments: number;
  totalPlayers: number;
  totalRegistrations: number;
}

export interface HistoryEntry {
  id: number;
  userId: number;
  tournamentId: number;
  tournamentName: string;
  action: string;
  createdAt: string;
}

export interface ScoreboardEntry {
  id: number;
  tournamentId: number;
  registrationId: number;
  squadName: string;
  kills: number;
  /** @nullable */
  rank?: number | null;
  points: number;
  updatedAt: string;
}

export interface ScoreboardUpdateBody {
  registrationId: number;
  squadName: string;
  kills: number;
  /** @nullable */
  rank?: number | null;
  points: number;
}

export type GetTournamentsParams = {
  status?: string;
  type?: string;
};

export type GetRegistrationsParams = {
  tournamentId?: number;
  userId?: number;
  status?: string;
};

export type GetResultsParams = {
  tournamentId?: number;
};