# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elite FF Tournament Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registrations:
    get:
      operationId: getRegistrations
      tags: [registrations]
      summary: Get registrations (filtered by tournament or user)
      parameters:
        - name: tournamentId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Registrations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Register for a tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegistrationBody"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{id}:
    get:
      operationId: getRegistration
      tags: [registrations]
      summary: Get registration details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{id}/verify:
    post:
      operationId: verifyRegistration
      tags: [registrations]
      summary: Verify a registration (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRegistrationBody"
      responses:
        "200":
          description: Registration verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{id}/decline:
    post:
      operationId: declineRegistration
      tags: [registrations]
      summary: Decline a registration (Host only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeclineRegistrationBody"
      responses:
        "200":
          description: Registration declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{id}/cancel:
    post:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel a registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
      required:
        - id
        - username
        - role
...
    Registration:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        userId:
          type: integer
        squadName:
          type: string
        playerNames:
          type: string
        paymentScreenshotUrl:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        utrNumber:
          type: string
        status:
          type: string
        slotNumber:
          type: ["integer", "null"]
        declineReason:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - tournamentId
        - userId
        - squadName
        - playerNames
        - utrNumber
        - status
        - createdAt
        - user
    CreateRegistrationBody:
      type: object
      properties:
        tournamentId:
          type: integer
        squadName:
          type: string
        playerNames:
          type: string
        paymentScreenshotUrl:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        utrNumber:
          type: string
      required:
        - tournamentId
        - squadName
        - playerNames
        - utrNumber
    VerifyRegistrationBody:
      type: object
      properties:
        slotNumber:
          type: ["integer", "null"]
      required: []
    DeclineRegistrationBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...